--- a/REQUIRE
+++ b/REQUIRE
@@ -1,3 +1,4 @@
 julia 0.4-
 Distributions
 SVMLightLoader
+Devectorize