--- a/src/SoftConfidenceWeighted.jl
+++ b/src/SoftConfidenceWeighted.jl
@@ -1,5 +1,7 @@
 module SoftConfidenceWeighted
 
+using Devectorize
+
 import Distributions: Normal, cdf
 import SVMLightLoader: SVMLightFile
 
@@ -53,12 +55,12 @@ normal_distribution = Normal(0, 1)
 
 
 function calc_margin{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
-    return label * dot(scw.weights, x)
+    @devec t = label .* dot(scw.weights, x)
 end
 
 
 function calc_confidence{T<:AA}(scw::SCW, x::T)
-    return dot(x, (scw.covariance .* x))
+    @devec t = dot(x, scw.covariance .* x)
 end
 
 
@@ -108,7 +110,7 @@ end
 
 
 function loss{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
-    t = label * dot(scw.weights, x)
+    @devec t = label .* dot(scw.weights, x)
     if t >= 1
         return 0
     end
@@ -130,36 +132,50 @@ function calc_beta{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
 end
 
 
-function update_covariance{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
+function update_covariance{S<:AA,T<:AA,R<:Real}(t::S, scw::SCW, x::T, label::R)
     beta = calc_beta(scw, x, label)
     c = scw.covariance
-    scw.covariance -= beta * (c .* x) .* (c .* x)
+    @devec t[:] = (c .* x) .* (c .* x)
+    BLAS.axpy!(-beta, t, scw.covariance)
     return scw
 end
 
 
-function update_weights{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
+function update_weights{S<:AA,T<:AA,R<:Real}(t::S, scw::SCW, x::T, label::R)
     alpha = calc_alpha(scw, x, label)
-    scw.weights += alpha * label * (scw.covariance .* x)
+    @devec t[:] = scw.covariance .* x
+    BLAS.axpy!(alpha * label, t, scw.weights)
+    return scw
+end
+
+
+function update{S<:AA,T<:AA,R<:Real}(t::S, scw::SCW, x::T, label::R)
+    x = vec(full(x))
+    if loss(scw, x, label) > 0
+        scw = update_weights(t, scw, x, label)
+        scw = update_covariance(t, scw, x, label)
+    end
     return scw
 end
 
 
 function update{T<:AA,R<:Real}(scw::SCW, x::T, label::R)
     x = vec(full(x))
+    t = Array(Float64, length(x))
     if loss(scw, x, label) > 0
-        scw = update_weights(scw, x, label)
-        scw = update_covariance(scw, x, label)
+        scw = update_weights(t, scw, x, label)
+        scw = update_covariance(t, scw, x, label)
     end
     return scw
 end
 
 
 function train{T<:AA,R<:AA}(scw::SCW, X::T, labels::R)
+    t = Array(Float64, size(X, 1))
     for i in 1:size(X, 2)
         label = labels[i]
         assert(label == 1 || label == -1)
-        scw = update(scw, slice(X, :, i), label)
+        scw = update(t, scw, slice(X, :, i), label)
     end
     return scw
 end