I did a few tweaks here and there such as using BLAS routines, devectorization and preallocation which reduce the memory footprint 100-fold. The runtime also reduces a little but I am not sure if it is statistically significant. 

However, this PR does introduce a new dependency on Devectorize.jl. Also, I am using the master branch of that Package right now, and not the last tagged version. (it gets rid of the deprecated warnings)

Here is the difference on your benchmark:

Old (without this PR)
```
======================= First run:
  7.244674 seconds (1.70 M allocations: 627.285 MB, 0.81% gc time)
======================= Second run:
  5.399670 seconds (652.07 k allocations: 408.219 MB, 0.30% gc time)
```

New (with this PR)
```
======================= First run:
  6.339811 seconds (533.85 k allocations: 16.442 MB)
======================= Second run:
  4.943052 seconds (245.50 k allocations: 4.982 MB, 0.28% gc time)
```