@IshitaTakeshi 2015-09-09
Thanks for awesome performance improvement. 
Now it is merged into master.